refactor: break down WhitelistBlockchainsCard
