+++ app/routes/dashboard.apps.$appId.security/view.tsx
@@ -34,6 +34,7 @@ import { EndpointQuery } from "~/models/portal/sdk"
 import { AmplitudeEvents, trackEvent } from "~/utils/analytics"
 import { getImageForChain } from "~/utils/known-chains/known-chains"
 import SecretKeyCard from "./components/SecretKeyCard/SecretKeyCard"
+import WhitelistBlockchainsCard from "./components/WhitelistBlockchainsCard"
 
 /* c8 ignore start */
 export const links = () => {
@@ -90,13 +91,6 @@ export const SecurityView = ({ endpoint, appId, blockchains }: SecurityViewProps
     whitelistMethodsDropdown,
   } = state
 
-  const addIfMissingOrRemove = (item: string, arr: string[]) => {
-    if (arr.indexOf(item) !== -1) {
-      return arr.filter((i) => i !== item)
-    }
-    return [...arr, item]
-  }
-
   const addIfMissing = (item: string, arr: string[]) => {
     if (arr.indexOf(item) !== -1) {
       return arr
@@ -116,94 +110,7 @@ export const SecurityView = ({ endpoint, appId, blockchains }: SecurityViewProps
       <securityAction.Form action={`/api/${appId}/settings`} method="post">
         <input name="appID" type="hidden" value={appId} />
         <SecretKeyCard endpoint={endpoint} />
-        <Card>
-          <div className="pokt-card-header">
-            <h3>{t.security.headings.approvedChains}</h3>
-            {isApprovedChainsSaveShown ? (
-              <Flex>
-                <Button
-                  mr=".5em"
-                  px="2em"
-                  size="sm"
-                  variant="outline"
-                  onClick={() => {
-                    dispatch({
-                      type: "SET_WHITELIST_BLOCKCHAINS",
-                      payload: endpoint.gatewaySettings.whitelistBlockchains as string[],
-                    })
-                    dispatch({
-                      type: "SET_SAVE_MODAL_SHOWN",
-                      payload: { modal: "isApprovedChainsSaveShown", shown: false },
-                    })
-                  }}
-                >
-                  {t.common.reset}
-                </Button>
-                <Button
-                  size="sm"
-                  sx={{
-                    padding: "0 2em",
-                  }}
-                  type="submit"
-                  variant="filled"
-                  onClick={() => {
-                    trackEvent(AmplitudeEvents.SecuritySettingsUpdate)
-                  }}
-                >
-                  {t.common.save}
-                  {navigation.state !== "idle" && (
-                    <Loader color={theme.colors.blue[5]} ml={8} size="xs" />
-                  )}
-                </Button>
-              </Flex>
-            ) : null}
-          </div>
-          <Flex justify="space-between">
-            <Text size="sm">{t.security.approvedChainsText}</Text>
-            <ChainsDropdown
-              chains={blockchains}
-              checkboxData={whitelistBlockchains}
-              onChange={(val: string) => {
-                dispatch({
-                  type: "SET_WHITELIST_BLOCKCHAINS",
-                  payload: addIfMissingOrRemove(val, whitelistBlockchains),
-                })
-                dispatch({
-                  type: "SET_SAVE_MODAL_SHOWN",
-                  payload: { modal: "isApprovedChainsSaveShown", shown: true },
-                })
-              }}
-            />
-          </Flex>
-
-          {whitelistBlockchains.map((item: string) => {
-            const blockchain: Blockchain | undefined | null = blockchains.find(
-              (c) => c?.id === item,
-            )
-            return (
-              <React.Fragment key={item}>
-                <AppEndpointUrl
-                  key={item}
-                  readOnly
-                  chain={blockchain}
-                  handleRemove={() => {
-                    const newArray = state.whitelistBlockchains.filter(
-                      (blockchain) => blockchain !== item,
-                    )
-                    dispatch({ type: "SET_WHITELIST_BLOCKCHAINS", payload: newArray })
-                    dispatch({
-                      type: "SET_SAVE_MODAL_SHOWN",
-                      payload: { modal: "isApprovedChainsSaveShown", shown: true },
-                    })
-                  }}
-                  hasDelete={true}
-                  value={blockchain?.description ?? ""}
-                />
-                <input name="whitelistBlockchains" type="hidden" value={item} />
-              </React.Fragment>
-            )
-          })}
-        </Card>
+        <WhitelistBlockchainsCard blockchains={blockchains} endpoint={endpoint} />
         <Card>
           <div className="pokt-card-header">
             <h3>{t.security.headings.userAgents}</h3>
